requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
cffi==2.0.0
contextlib2==21.6.0
cryptography==46.0.3
Django==5.2.5
django-ninja==1.5.1
django-ninja-extra==0.30.8
django-ninja-jwt==5.4.2
injector==0.23.0
pillow==11.3.0
psycopg2==2.9.10
pycparser==2.23
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
PyJWT==2.10.1
python-dotenv==1.1.1
sqlparse==0.5.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
